Extracting options from lmdb::env
