+++ nano/store/lmdb/lmdb.cpp
@@ -41,7 +41,7 @@ nano::store::lmdb::component::component (nano::logger & logger_a, std::filesyste
 	version_store{ *this },
 	rep_weight_store{ *this },
 	logger{ logger_a },
-	env (error, path_a, nano::store::lmdb::env::options::make ().set_config (lmdb_config_a).set_use_no_mem_init (true)),
+	env (error, path_a, nano::store::lmdb::options::make ().set_config (lmdb_config_a).set_use_no_mem_init (true)),
 	mdb_txn_tracker (logger_a, txn_tracking_config_a, block_processor_batch_max_time_a),
 	txn_tracking_enabled (txn_tracking_config_a.enable)
 {
@@ -127,7 +127,7 @@ bool nano::store::lmdb::component::vacuum_after_upgrade (std::filesystem::path c
 		std::filesystem::rename (vacuum_path, path_a);
 
 		// Set up the environment again
-		auto options = nano::store::lmdb::env::options::make ()
+		auto options = nano::store::lmdb::options::make ()
 					   .set_config (lmdb_config_a)
 					   .set_use_no_mem_init (true);
 		env.init (error, path_a, options);
